+++ schema/class-schema.php
@@ -25,6 +25,11 @@ class Schema {
 
 
 		/**
+		 * Single Post Loaded Action
+		 */
+		add_action( 'template_redirect', [ $this, 'maybe_build_single_cpt_schema' ] );
+
+		/**
 		 * Blokki Loop action
 		 */
 		add_action( 'blokki_template_posts_loop_post', [ $this, 'setup_schema_for_post_in_loop' ], 10, 3 );
@@ -53,6 +58,30 @@ class Schema {
 	public function is_active_schema_support() {
 
 		return function_exists( 'get_field' ) && get_field( 'blokki_schema_support', 'options' );
+
+	}
+
+	/**
+	 * Maybe build schema for single cpt
+	 * @hooked template_redirect
+	 */
+	public function maybe_build_single_cpt_schema() {
+
+		if ( ! is_singular() ) {
+			return;
+		}
+
+		if ( ! $this->is_active_schema_support() ) {
+			return;
+		}
+
+		global $post;
+
+		// we are only interested in WP_Post Object
+		if ( 'WP_Post' === get_class( $post ) ) {
+			$this->enqueue_post_single_schema( $post );
+		}
+
 
 	}
 
@@ -94,6 +123,36 @@ class Schema {
 	/**
 	 * Enqueue Schema using the $post object
 	 */
+	public function enqueue_post_single_schema( $post ) {
+		$post = get_post( $post );
+
+		if ( ! $post ) {
+			return null;
+		}
+
+		$single_schema_type = $this->get_post_single_schema_type( $post );
+
+		if ( ! $single_schema_type ) {
+			return null;
+		}
+
+		/**
+		 * Lets setup Schema Type class for the specified loop_schema
+		 */
+		$single_schema_type_class = $this->setup_schema_type( $single_schema_type, $post->post_type );
+
+		if ( ! $single_schema_type_class ) {
+			return null;
+		}
+		/**
+		 * Add Post Schema to the class
+		 */
+		$single_schema_type_class->add_post_schema( $post );
+	}
+
+	/**
+	 * Enqueue Schema using the $post object
+	 */
 	public function enqueue_post_loop_schema( $post ) {
 		$post = get_post( $post );
 
@@ -127,6 +186,15 @@ class Schema {
 	public function get_post_loop_schema_type( $post ) {
 
 		return $this->get_post_schema_type( $post, true );
+
+	}
+
+	/**
+	 * Get Loop Schema Type from Post Object
+	 */
+	public function get_post_single_schema_type( $post ) {
+
+		return $this->get_post_schema_type( $post, false );
 
 	}